pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperclaude"
version = "0.1.0"
description = "Swarm orchestration for Claude Code - coordinate multiple Claude instances"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
hyperclaude = "hyperclaude.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
hyperclaude = ["templates/*.md"]
